Add bifrost manifest
Add class to clone bifrost from source, add config file templates and
run the install procedure.

The config templates are currently highly tailored to Infra's use case
and may need to be reworked if we wish to make this consumable
downstream.

Based on the ironic::bifrost class[1] but with some modifications to
avoid heavy reliance on the puppet DSL.

Co-Authored-By: Clint Byrum <clint@fewbar.com>
Co-Authored-By: greghaynes <greg@greghaynes.net>
Co-Authored-By: Ricardo Carrillo Cruz <ricardo.carrillo.cruz@gmail.com>

[1] http://git.openstack.org/cgit/openstack/puppet-ironic/tree/manifests/bifrost.pp

Change-Id: I949344c16cf9ee3965b0bc96850eb208ac65b168