Set filesystem_store_datadir explicitly in glance
In kilo, the glance_store/filesystem_store_datadir defaults to
/var/lib/glance/images in glance-api.conf[1]. In liberty, it has no
default[2]. This patch uses the glance::backend::file class to set this
value and to make it explicit in our puppet code that we are using the
file backend. This will be a no-op on the current kilo deployment but
will facilitate upgrading to liberty.

[1] http://docs.openstack.org/kilo/config-reference/content/section_glance-api.conf.html
[2] http://docs.openstack.org/liberty/config-reference/content/section_glance-api.conf.html

Change-Id: I6e8cf1a823699975a607144c691bd03e71000089