Use SSL for rabbitmq traffic
With rabbitmq traffic now encrypted, we can use the public internet
instead of a private network for rabbitmq. This also means we don't
need to specify a management address any more since rabbitmq was the
only service using it.

This patch includes a workaround to make sure the new neutron config
causes a refresh of the neutron-plugin-linuxbridge-agent service, which
will otherwise not be able to talk on rabbitmq and causes nova boots to
fail.

Assuming all goes well this will cause a brief downtime, so nodepool
should be temporarily disabled when applying this.

Change-Id: I5a25a5e4aa70db66db2d9331d7f5e4ac8b785002