Update SSL configuration
We don't need to use a separate CA chain for these certs, instead it is
sufficient to just trust the self-signed cert.

Moreover, we have been cheating by copying the same certs and keys to
various directories for each service, so instead of bothering with
keeping separate certs let's just formalize having a single pair. The
cert will be used as its own CA and added to the system trusted
certificates. The key still needs to be privately readable by certain
system users, so we'll still copy that into the ssl directories for
each service.

Also, since we'll be changing these keys, make sure they are set up to
notify the service they're supporting.

Additionally, automate the trusting of our self-signed certs on the
compute hosts which previously was done manually. The compute hosts
need to be able to use encrypted rabbitmq and make API calls to
keystone and neutron.

Change-Id: Ibeea608e965e58c496a95b2f02a4bf6b13e15f0e