Use unversioned identity endpoints
In Liberty with Keystone v3 we don't need to include the 'v2.0' or 'v3'
string in the endpoint. In the next release of the keystone puppet
module, the endpoint will default to unversioned. To clean up our
endpoints and avoid a deprecation message, set the version explicitly
to empty string.

Change-Id: I893393ee165757aa5f4581b144009528e02806df