Reconcile compute and controller classes
The infracloud::controller and infracloud::compute classes contain
several common classes, but we're configuring them slightly differently.
This is a problem for someone who wants to run an "all-in-one" type of
scenario where the compute is applied on top of the controller, because
puppet will overwrite settings that are different between the two types
of nodes. This patch makes what can be common identical. It also pulls
out the mysql string setting into a separate class so that the compute
node does not need access to the mysql password. This patch should be
entirely a no-op change but will prepare us to add an all-in-one
functional test suite in a followup patch.

Change-Id: Ie1d91b642f45f8b742b6ac0f9d4f1981e63ded9f