Parameterize libvirt virt type
This change allows us to control the virtualization technology running
under nova. In production, we will always use KVM as described in the
infra-cloud documentation[1]. However, in functional tests (coming in a
followup change) we'll need to change it to qemu since we don't have
nested virtualization[2]. This is technically volkswagoning but this
will still be effective for testing the majority of our puppet changes.

[1] http://docs.openstack.org/infra/system-config/infra-cloud.html#architecture
[2] http://logs.openstack.org/40/288840/4/check/gate-puppet-infracloud-puppet-beaker-rspec-dsvm-trusty-nv/6f0501c/logs/nova/nova-compute.txt.gz

Change-Id: I85b739763cde9610a2bf09090d4ca76a32c58a51