Enable inventory_dhcp_static_ip on dnsmasq
In the past, we had issues in West due to rogue machines on the vlan
we were given grabbing DHCP leases, leaving legit servers from the
infracloud to be provisioned as it the leases pool exhausted.
We added the inventory_dhcp_static_ip flag to have dnsmasq assigning
a fixed set of IPs to the servers provisioned. This then allowed
dnsmasq to filter which machines it would hand out a lease for.

Change-Id: I40916b18c86781e5ae294d4f25ccf4214de09759