Switch to infra-cloud-bridge element
The infra-cloud-networking was the way to work on the past HP East,
where the public IP range was served via a double DHCP on baremetal
node.
We put changes in Bifrost that eventually allowed us to setup the
provisioning IP addresses and the public IPs are then set on the
config drives.

Since Glean is not able to configure bridges, this element does that
for us, by consuming the network_info data from the config drive.

Change-Id: Ie1b2872a0b4fe84289aaddc8d129caaa9e0166d9