Search for VLAN devices on config drive to create bridge on it
Currently, the infra-cloud-bridge element expects the VLAN device
being the first interface on the config drive links section.
That used to work, but in my testing it seems that due to some
changes in Glean that's no longer the case and the VLAN is after
the physical device.
Try to be more smart by searching the VLAN device if exists and create
the bridge on it, otherwise just fallback to the physical device.

Change-Id: I028ce200c0cc7d30b2ebb3a159d26005a06021f0