Add the ability to configure bridge name
Naming with dashes is causing issues on CentOS,
because glean is called with interface br/xxx instead
of br-xxx due some naming translation.
Also interface name and vlans can differ on environments,
causing all puppet modules to change their naming. Instead of
depending on vlan or nic to configure bridge name, add the
ability to pass that name as a parameter, and likely set
to br_infracloud on our deployments.
At the moment, leaving the default naming to do not break
existing cloud deployment.

Change-Id: I0790bab2fd63f525e8b9c8d47ee79ea63a72750a