Remove per-region handling
Since vanilla and chocolate are on the same VLAN, we don't have
two different bifrost machines, just one to handle both.
Therefore, the per-region split logic is not needed anymore and makes
it impossible to have on the current baremetal machine in prod
having both inventories at the same time.

Change-Id: I433cdb8675c3acd2e699272aaf6a0e67eb9200b0