Get rid of ipv4- prefix when looking for sys-interfaces
bifrost is composing network ids with ipv4+mac, while
sys_interfaces is just a dict with mac:nic values.
Get rid of that ipv4 prefix when searching for the right
interface to create the bridge.
We will be trying first to retrieve the original id,
then get rid of ipv4- as a fallback.

Change-Id: I78ac1e6b735cbc2cc120c92f6c9bf1492924f5e4