Fix beaker on xenial
This patch adds a xenial nodeset and updates the spec helper to install
puppet 3 from the Ubuntu repos instead of from puppetlabs. However, it
ultimately disables the tests on xenial because neither the Ubuntu Cloud
Archive[1] nor the Puppet-OpenStack modules support mitaka on xenial. We
will need to plan our xenial upgrade and OpenStack upgrade at the same
time and reenable xenial testing when we are ready for the OpenStack
upgrade.

[1] http://ubuntu-cloud.archive.canonical.com/ubuntu/dists/xenial-updates/

Change-Id: I3b7c2c83d6a20907de4b69fe6fcd4c59a5052d23